feat(traits): add Validate trait
Introduce a new Validate trait for types that can be validated as a TypeID prefix. This trait is automatically implemented for any type that implements AsRef<str> and can be converted to a TypeIdPrefix using TryFrom.

NEW FEATURE: You appear to have introduced one or more new features that are backward-compatible. According to Semantic Versioning (SemVer), this requires a minor version update. Please verify and update your version number accordingly.